requirements.txt
fastapi==0.109.2
numpy==1.26.4
pandas==2.0.3
pillow==10.2.0
pydantic==2.5.3
pydantic-core==2.14.6
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
torch==2.1.0
torchvision==0.16.0
uvicorn==0.27.1